Adj. filenames; function to combine conversations
We can now extract all conversations from a PDF. There is only one more
issue. Conversations are combined by page, so dialogue going less than a
page or longer will be cut off or combined with other conversation. This
is not optimal but it is close to impossible to define a generic cutoff
value here.